+++ src/openapi/helpers/types.ts
